+++ app/Http/Controllers/Admin/MediaCenter/InspectionCertificateController.php
@@ -406,13 +406,15 @@ class InspectionCertificateController extends Controller
                     }
 
                     // 6) Save DB record
-                    $media                 = Media::findOrNew($post->mediaOne->id ?? null);
-                    if($post->mediaOne != null){
-                        if (Storage::disk('images')->exists($media->filepath)) {
+                    $oldMediaId = $post->mediaOne->id ?? null;
+                    $media = Media::findOrNew($oldMediaId);
+                    
+                    if($oldMediaId){
+                        if ($media->filepath && Storage::disk('images')->exists($media->filepath)) {
                             Storage::disk('images')->delete($media->filepath);
                         }
                         // Delete the thumbnail file from storage
-                        if (Storage::disk('images')->exists($media->thumbnailpath)) {
+                        if ($media->thumbnailpath && Storage::disk('images')->exists($media->thumbnailpath)) {
                             Storage::disk('images')->delete($media->thumbnailpath);
                         }
                     }
